Code cleanup
- Used "<Link>" component instead of an "<a>" element since that works faster and doesn't cause a page refresh.
- Added a confirmation message when deleting an account.
- Redirected to the home page when deleting your account.

+++ frontend/src/Account.jsx
@@ -2,9 +2,11 @@ import { useState } from "react";
 import Alert from "react-bootstrap/Alert";
 import Container from "react-bootstrap/Container";
 import Button from "react-bootstrap/Button";
+import { Link, useHistory } from "react-router-dom";
 import { ROUTES } from "./constants";
 
 export function Account({ username, onDelete }) {
+  const history = useHistory();
   const [disableDeleteAccount, setDisableDeleteAccount] = useState(false);
   const [deleteAccountError, setDeleteAccountError] = useState("");
 
@@ -15,13 +17,13 @@ export function Account({ username, onDelete }) {
 
       <hr />
 
-      <a href={ROUTES.EDIT_PASSWORD}>
+      <Link to={ROUTES.EDIT_PASSWORD}>
         <p>Edit Password</p>
-      </a>
+      </Link>
 
-      <a href={ROUTES.MANAGE_SOCIAL_PROFILES}>
+      <Link to={ROUTES.MANAGE_SOCIAL_PROFILES}>
         <p>Manage Social Media Profiles</p>
-      </a>
+      </Link>
 
       <hr />
 
@@ -29,14 +31,25 @@ export function Account({ username, onDelete }) {
         variant="danger"
         disabled={disableDeleteAccount}
         onClick={async () => {
+          if (
+            !window.confirm(
+              "Are you sure you want to delete your account and all associated data? " +
+                "This cannot be undone."
+            )
+          ) {
+            return;
+          }
+
           setDisableDeleteAccount(true);
 
           try {
             await onDelete(username);
-          } catch (e) {
-            setDeleteAccountError(e.message);
-          } finally {
             setDisableDeleteAccount(false);
+
+            history.push(ROUTES.HOME);
+          } catch (e) {
+            setDisableDeleteAccount(false);
+            setDeleteAccountError(e.message);
           }
         }}
       >
